requirements.txt
fastapi==0.100.1
uvicorn==0.25.0
requests==2.31.0
cachetools==5.3.1
matplotlib==3.9.2
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.0